updated routing to profile page and its content
