+++ src/components/login/Login.js
@@ -180,7 +180,6 @@ class Login extends React.Component {
             <Label>Password</Label>
             <InputField
                 type="password"
-
                 placeholder="Enter here.."
                 onChange={e => {
                   this.handleInputChange("password", e.target.value);
