+++ src/components/login/Registration.js
@@ -2,10 +2,8 @@ import React from "react";
 import styled from "styled-components";
 import { BaseContainer } from "../../helpers/layout";
 import { getDomain } from "../../helpers/getDomain";
-import User from "../shared/models/User";
-import {Redirect, withRouter} from "react-router-dom";
+import { withRouter } from "react-router-dom";
 import { Button } from "../../views/design/Button";
-import {Promise as resolve} from "q";
 
 const FormContainer = styled.div`
   margin-top: 2em;
@@ -67,7 +65,6 @@ class Registrator extends React.Component {
             username: null,
             password: null
         };
-        this.redirect = null; //used for redirectLogin() only
         this.today = new Date();
     }
 
@@ -84,63 +81,40 @@ class Registrator extends React.Component {
     }
 
     register() {
-        var existingUser = new User();
-        //var date = new Date();
-        //var timestamp = date.getTime();
-
-        fetch(`${getDomain()}/users/${this.state.username}`, { //check for existing user
-            method: "GET",
+        //alert("Got to register()");
+        fetch(`${getDomain()}/users/`, { //try registering user
+            method: "POST",
             headers: {
                 "Content-Type": "application/json"
             },
             body: JSON.stringify({
-                username: this.state.username,
-                password: this.state.password
+                firstName: this.state.firstName,
+                lastName: this.state.lastName,
+                birthDate: this.state.birthdate,
+                password: this.state.password,
+                username: this.state.username
             })
         })
-            .then(response => response.json())
-            .then(returnedUser => {
-                existingUser = new User(returnedUser);
-                //returnedUser.print();
-                if(existingUser.username !== this.state.username || existingUser.password !== this.state.password) {//if inexistent register new user
-
-                    //NON-FUNCTIONAL CHECK FOR EXISTING USER!!
-
-                    //alert("Got to second fetch");
-                    fetch(`${getDomain()}/users`, {
-                        method: "POST",
-                        headers: {
-                            "Content-Type": "application/json"
-                        },
-                        body: JSON.stringify({
-                            firstName: this.state.firstName,
-                            lastName: this.state.lastName,
-                            birthDate: this.state.birthdate,
-                            password: this.state.password,
-                            username: this.state.username
-                        })
-                    })
-                        .then(() => {
-                            this.redirectLogin();
-                        })
-                        .then( () => {
-                            alert("Registration successful. Try logging in with your new user credentials");
-                        })
-                    .catch(err => {//error during 2nd fetch
-                        if (err.message.match(/Failed to fetch/)) {
-                            alert("The server cannot be reached. Did you start it?");
-                        /**} else if(existingUser.username == null) {
-                            alert("User not existing. You should register first");*/
-                        } else {
-                            alert(`Something went wrong during the login: ${err.message}`)
-                            //TODO: catch errors for user already existing
-                            //TODO: redirect to register screen in case of error
-                        }
-                    });
+            .then((response, err) => {
+                if(response.status === 409) {
+                    console.log(`ERROR: Failed to register user ${this.state.username} with status 409 CONFLICT`);
+                    if (err.message.match(/Failed to fetch/)) {
+                        alert("The server cannot be reached. Did you start it?");
+                    } else {
+                        alert(`Something went wrong during the login: ${err.message}`)
+                        //TODO: catch errors for user already existing
+                        //TODO: redirect to register screen in case of error
+                    }
                 } else {
-                    alert("User already existing");
+                    console.log(`OK: Successfully registered user ${this.state.username}`);
                 }
-            });
+            })
+            .then(() =>  {
+                this.redirectLogin();
+            })
+            .then(() => {
+                alert("Registration successful. Try logging in with your new user credentials");
+            })
     }
 
     render() {
